ln02/requirements.txt
markdown-it-py==2.2.0
mdurl==0.1.2
psycopg2-binary==2.9.5
Pygments==2.14.0
rich==13.3.3
